separated out article submit and file upload on researcher page, removed doclink from article model
