+++ client/src/pages/researcher.js
@@ -1,9 +1,7 @@
-import React, { useState, useEffect, useRef } from "react";
+import React, { useState, useEffect} from "react";
 
 import MyDropzone from "../components/dragAndDrop";
 import axios from "axios";
-
-
 import API from "../utils/articleAPI";
 import FilesList from "../components/filesList"
 
@@ -13,13 +11,6 @@ import FilesList from "../components/filesList"
 function Researcher() {
   // set component's initial state
   const [formObject, setFormObject] = useState({});
-  const [file, setFile] = useState(null);
-  const [state, setState] = useState({
-    title: "",
-    description: ""
-  });
-  const [errorMsg, setErrorMsg] = useState("");
-  const dropRef = useRef();
 
   //update component state when the user types into input field
   function handleInputChange(event) {
@@ -27,19 +18,17 @@ function Researcher() {
     setFormObject({ ...formObject, [name]: value });
   }
 
-  //handle submit for the file drop
-  const handleOnSubmit = async (event) => {
+  //handle submit for article collection
+  const handleFormSubmit= async (event) => {
     event.preventDefault();
     //creates article
     if (
-      formObject.researchDocLink &&
       formObject.articleName &&
       formObject.authorName &&
       formObject.articleAbstract
     ) {
       API
         .saveArticle({
-          researchDocLink: formObject.researchDocLink,
           articleName: formObject.articleName,
           authorName: formObject.authorName,
           articleAbstract: formObject.articleAbstract,
@@ -47,40 +36,9 @@ function Researcher() {
         .then(console.log(formObject))
         .catch((err) => console.log(err));
     }
-    //creates file
-    try {
-      const { title, description } = state;
-      if (title.trim() !== "" && description.trim() !== "") {
-        console.log(file);
-        if (file) {
-          const formData = new FormData();
-          formData.append("file", file);
-          formData.append("title", title);
-          formData.append("description", description);
-
-          setErrorMsg("");
-          await axios.post(`/api/files/upload`, formData, {
-            headers: {
-              "Content-Type": "multipart/form-data"
-            }
-          });
-        } else {
-          setErrorMsg("Please select a file to add.");
-        }
-      } else {
-        setErrorMsg("Please enter all field values");
-      }
-    } catch (error) {
-      error.response && setErrorMsg(error.response.data);
-    }
   };
 
 
-  //handle form submit uses API.saveArticle method to save data
-  // function handleFormSubmit(event) {
-  //   event.preventDefault();
-    
-  // }
   return (
     <div className="container">
       <div className="row">
@@ -88,13 +46,28 @@ function Researcher() {
           <div className="jumbotron">
             <div className="container">
               <h2>Directions for Researchers</h2>
+              <br/>
+              <p>Welcome to the Educational Data Stream, where your research can impact K-12 students across the United States! 
+                Please first complete the Article Information with the Article Name, Author Name, and a brief abstract. Then, upload your Article Information. </p>
+                <p>
+                Then, upload your file. You can then drag and drop your file in the dropzone, or click to browse. 
+                Enter a name for the file to be saved as and a brief description. 
+                </p>
+                <p>
+                After you submit, our editors will curate content for educators to 
+                access and put your research to work shaping future generations.
+              </p>
+              <p>
+                Thank you! 
+              </p>
+              <p>-The EDS Team-</p>
               
             </div>
           </div>
         </div>
         <div className="col">
           <div className="jumbotron">
-            <h2>Upload Form</h2>
+            <h2>Article Information</h2>
             <div className="form-group">
               <input
                 className="form-control"
@@ -120,16 +93,15 @@ function Researcher() {
                 placeholder="Abstract (required)"
               ></textarea>
             </div>
-            <div className="form-conrol mb-4">
-            <MyDropzone />
-            </div>
+            
             <button
               type="button"
               className="btn btn-secondary"
-              onClick={handleOnSubmit}
+              onClick={handleFormSubmit}
             >
-              Submit
+              Submit Article Information
             </button>
+            <MyDropzone />
           </div>
         </div>
       </div>
